ref(prettier): move husky rules to own directory
