+++ src/prettier/index.ts
@@ -1,14 +1,7 @@
-import {
-  apply,
-  chain,
-  mergeWith,
-  noop,
-  Rule,
-  Tree,
-  url
-} from '@angular-devkit/schematics';
+import { chain, noop, Rule } from '@angular-devkit/schematics';
 import { installDependencies } from '../lib';
 import { PrettierSchematicOptions } from './models';
+import { applyHuskyConfiguration } from './rules/husky';
 import {
   configurePrettier,
   registerPrettier,
@@ -31,32 +24,6 @@ export function configureHusky(): Rule {
     installDependencies({
       devDependencies: ['husky', 'pretty-quick', 'lint-staged']
     }),
-    setupConfigurationFiles()
+    applyHuskyConfiguration()
   ]);
-}
-
-export function setupConfigurationFiles(): Rule {
-  return (tree: Tree) =>
-    tree.exists('.huskyrc.json')
-      ? updateHuskyRc(tree.read('.huskyrc.json')!)
-      : createHuskyRc();
-}
-
-export function createHuskyRc(): Rule {
-  return mergeWith(apply(url('./templates/hook'), []));
-}
-
-export function updateHuskyRc(origin: Buffer): Rule {
-  return (tree: Tree) => {
-    const originHuskyRc = JSON.parse(origin.toString('utf-8'));
-    const updatedHuskyRc = {
-      ...originHuskyRc,
-      hooks: {
-        ...originHuskyRc.hooks,
-        'pre-commit': 'pretty-quick --staged && lint-staged'
-      }
-    };
-
-    tree.overwrite('.huskyrc.json', JSON.stringify(updatedHuskyRc, null, 2));
-  };
 }
