+++ src/prettier/rules/husky/apply-husky-configuration.ts
@@ -1,0 +1,10 @@
+import { Rule, Tree } from '@angular-devkit/schematics';
+import { createHuskyRc } from './create-husky-rc';
+import { updateHuskyRc } from './update-husky-rc';
+
+export function applyHuskyConfiguration(): Rule {
+  return (tree: Tree) =>
+    !tree.exists('.huskyrc.json')
+      ? createHuskyRc()
+      : updateHuskyRc(tree.read('.huskyrc.json')!);
+}
