+++ src/prettier/rules/husky/create-husky-rc.ts
@@ -1,0 +1,4 @@
+import { apply, mergeWith, Rule, url } from '@angular-devkit/schematics';
+export function createHuskyRc(): Rule {
+  return mergeWith(apply(url('./templates/hook'), []));
+}
