+++ src/prettier/rules/husky/index.ts
@@ -1,0 +1,1 @@
+export * from './apply-husky-configuration';
